pass-import chech for decrypt is Ok

+++ config/fish/functions/pass-import.fish
@@ -3,11 +3,15 @@
 function pass-import --description 'import pass passwords'
 	if test -f pass.tar.gz.gpg
 		gpg --decrypt pass.tar.gz.gpg > pass.tar.gz
-		tar xzvf pass.tar.gz
-		rm -rfv ~/.password-store/
-		mv -v .password-store/ ~/
-		rm -rfv pass.tar.*
-		echo "Passwords imported"
+		if test $status -eq 0
+			tar xzvf pass.tar.gz
+			rm -rfv ~/.password-store/
+			mv -v .password-store/ ~/
+			rm -rfv pass.tar.*
+			echo "Passwords imported"
+		else
+			echo "Invalid decryption"
+		end
 	else
 		echo "No import file"
 	end
